requirements.txt
asn1crypto==0.24.0
cffi==1.13.2
cryptography==2.8
Django==2.2.3
django-cors-headers==3.0.2
django-extensions==2.2.1
djangorestframework==3.10.2
Markdown==3.1.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
pydotplus==2.0.2
pyparsing==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
